Fix flicker on scratch sidebar creation.
This is done by setting the visibility to 'hidden' while the
back-and-forth between the application and scratch is not finished.

Setting the display to 'none' (instead of the visibility) wouldn't work
as if would affect the positioning/sizing of the element. The result of
this is that the elements (blocks) won't be visible event after the
workspace is shown back.

[Gitea] Closes #8.