Add base change.
This is the path, but it's not sufficient. As sometimes mnesia becomes
corrupted.

In some of these cases, stopping the process and starting it again will
be enought to fix it.
On other cases, removing the LATEST.LOG file will be needed for booting
up again. Take into account that if that file takes more than 97 bytes,
some of the latest transactions might be lost.