Remove tag autocompletion.
The reason for this is that the auto-completion only worked with already
present tags. As there's no reason for tags to be multiple times, they
should be deduplicated.

Also, for this reason, there's no point in facilitating writing them
twice. For this reason, and because that could be confusing, the
autocompletion functionality is removed for now.

Hopefully it can come in handy later, at which point it's possible to
revert this single commit.