Add UI and structure for custom Signal creation.
This is based on the idea that custom signals, like templates, which
their structure is based on, are not bound to a specific program, but to
the user that creates it.

This might is useful, as it allows a signal to be used by multiple
programs, and can be shared as any other kind of resource.