Workaround: Fix filter problem in Safari.
The problematic filter that made the blocks being dragged disappear in
Safari is replaced by a standard drop-shadow, which also doesn't work on
Safari, but allows the block to remain visible. The downside is that
Chrome also doesn't support drop-shadows, so it loses the usual filter
(though it keeps working).