Fix traduction of mouse position when hovering element with filter.
MoveMove(ev) event on Firefox reported an invalid ev.target when hovering an
element with a filter (as the node <rect> body). This happened when drawing the
line to connect IO ports between blocks.

The workaround for this is to add a '.drawing' class to the <svg> canvas while
line is being drawn by the user, and set the element with filters to
'pointer-events: none', so they are transparent to mouse events.