Add basic state management to workspace.
Take not of the current state of the workspace.
Possible states are 'waiting', 'dragging-block', 'dragging-workspace'.
Some events might be ignored in different states to avoid problems when
intermediate actions are performed outside the canvas.

This allows the UI to recover, for example, from dragging toolbox a block
outside the canvas, doing MouseUp. With this, the user can MouseDown on the
block in the canvas and everything will go back to normal.