WIP: Convert execution fork in a normal operation.
This means that a single pulse output cannot go to two different blocks without
a fork block in the middle to do the split.

If this becomes too "verbose", the placement of the block can be made automatic,
so it shouldn't become a problem later on. Making this automatic but visible
should improve the comprehensibility of the flows.