Restructure control flow to avoid glitches.
The solution is to check that each getter operation is only performed the first
time it is referred on each thread. In case it's referred twice in the first
block, add a preload block before it so it can be evaluated before.

Glitches: https://en.wikipedia.org/wiki/Reactive_programming#Glitches