Use jstz to detect browser timezone and offer it first.
This might not always yield a perfect result, the browser does not provide the
timezone itself. But it should be a reasonable default for most users. As such
it's just offered top of the list (and as default), and the user will be able to
modify it.

Later it might be interesting to allow the users to configure a default timezone
themselves, instead of relying on heuristics.