Implement authentication check.
Remove the deprecated `service_id` entry on #service_port_entry{}, and add
`old_skip_authentication`. The role of the new entry is to allow old bridges to
continue operating while they are not updated to use authentication tokens.

When a bridge connects to the websocket endpoint, it will be considered not
`authenticated` by default. When the first message is received on the server it will
check if it's a valid authentication message, and if not it will disconnect the
bridge.

The previous step will be required for the pre-existing bridges until a token is
used by the first time to authenticate them. After that it will be necessary everytime.