Prepare frontend for SSR through angular-universal.
Server-Side-Rendering allows for the pages to be rendered on the server, and
then transferred to the browser to continue it's usage (everything through the
API). It provides advantages in SEO (browser will get to read the main page) and
first content render, in exchange for additional server load and potentially
more network load.

Full Client-Side-Rendering is still the default operation method, so unless this
is a desired change, it can be ignored.