WIP: Establish connections directly on the program-building screen.
The goal of this is to avoid confusion when a user adds a bridge (or gets access
to one), goes to the program-building screen and doesn't find the bridge's
blocks.

In the new flow a user would add a bridge, or get access to one, and go directly
to the program-building screen. Here the bridges would be shown alongside other
blocks on the sidebar, but instead of showing the bridge's blocks, the editor
will show a button to establish the connection to the bridge. From there on, the
usual steps will be followed.

Note that:
 - In this commit only direct connections (the ones not requiring registration data) are
supported.
 - The connections "header" on the dashboard is removed, to avoid confusion with
 the bridge's tab.
 - Minimal (backwards compatible) changes to the backend API are needed.
 - No changes to the bridges are needed.