Remove unused automate_service_port_channel_table.
This information is now stored on the bridge's connection along with relevant
metadata.

Methods that referenced the old table have been removed, when possible, or
updated to use the connections table.

Storage of a bridge's last monitor value has been updated to be indexed by
bridge instead of by channel. This will avoid problems when a user has multiple
connections to a single bridge, and the data can come from any of them.