Prepare to use @ngx-utils/cookies as backend for SSR cookies.
This will be used to store authentication data. The reason for this is to make
this data available to the server during the SSR so it can be used to render
values that come from API queries.

This uses temporarily a fork of @ngx-utils/cookies, while the support needed for
Angular 10 is not merged. The associated PR for this support is
https://github.com/ngx-utils/cookies/pull/24.