Fix block overlapping problems.
SVG has a kind z-index where the "later" elements on the tree are the ones that
are shown over the previous ones. Before this commit this meant that the blocks
added later to the canvas where always placed above the ones before. Consider
the following process:

 - Add a button to the canvas
 - Add an horizontal section

As the horizontal section was added after the button, there would be no way to
place the button above the horizontal section.

This commit fixes this by moving the elements that should be on top to the end
of the canvas tree.

This works extends the mechanism used during the load, which loaded the elements
in a way that dependencies where loaded before dependent blocks. For example
first "Horizontal section" blocks are loaded, and later the "Button" blocks that
are contained inside them.

The way this commit fixes the canvas after the load has happened is by moving
the blocks that the user selects to the end of the tree, but moving before the
container elements and later the contained elements, thus making the contained
elements the latest ones.

This still has some quirks, but they will not be relevant as in the future the
overlapping won't happen unless in the case of container-contained groups.