Add function to reposition elements iteratively.
This is a temporary utility for testing, and is invoked by calling
`repositionIt()` on the browser terminal.

This repeats the `repositionAll` (also callable from the terminal with
`reposition()`) until a stable configuration is found or 100 iterations have
run.

This function runs each iteration with a 100ms separation, to avoid totally
stalling the browser.