Remove autoposition toggle, always set it to true.
Now that other resizing options, autopositioning is the only way to keep a
program tidy.

Although having a toggle that could be used to workaround cases
where the autopositioning is still buggy was interesting, disabling
autpositioning will make the programs get unwieldy.

For now requiring to be an
advanced user for FlowEditor programs should avoid the bigger problems in case
of positioning bugs.

This commit reverts:

 - Add autoposition setting as a slide toggle.

 - Move reposition toggle to advanced, auto-enable in some cases.

This reverts commit a700f095fd978ee3e672eb13a1e4c24a130f4252.