Implement jump-to-position on backend.
For this to work correctly with IF blocks, a new property has been added to
the #program_thread{}, the `direction`. Normally is set to `forward`, but when
"exiting" an IF or section it is set to `up`. This enables the IF blocks to take
decisions on where the execution comes from, from the block "before" of from the
one "downstream" from it.

With this, the IF blocks can be executed after a go-to and it won't be
misinterpreted as coming back from the IF contents.