Full restart secondary nodes when the primary is back up.
To do this track secondary nodes that half-restarted (which `automate_storage`
has restarted but the rest of the node hasn't), and trigger a full node restart.
With this we avoid having to do more complex synchronization and can keep the
service working at all times.

Consider a restarting primary node
 - At first the secondaries keep working
 - When the primary is back up again, the secondaries's `automate_storage`
 restart, getting the primary working again.
 - Now that the primary can answer the queries the secondaries are restarted to
 synchronize them.